ios build success

+++ driver_ev/src/net_ev_dgram.c
@@ -199,7 +199,7 @@ int net_ev_dgram_send(net_dgram_t base_dgram, net_address_t target, void const *
         }
     }
     
-    int nret = cpe_sendto(dgram->m_fd, data, data_len, CPE_SOCKET_DEFAULT_SEND_FLAGS, (struct sockaddr *)&addr, addr_len);
+    int nret = (int)cpe_sendto(dgram->m_fd, data, data_len, CPE_SOCKET_DEFAULT_SEND_FLAGS, (struct sockaddr *)&addr, addr_len);
     if (nret < 0) {
         CPE_ERROR(
             net_schedule_em(schedule), "ev: dgram: send %d data to %s fail, errno=%d (%s)",
@@ -229,7 +229,7 @@ static void net_ev_dgram_receive_cb(EV_P_ ev_io *w, int revents) {
     socklen_t addr_len = sizeof(addr);
     char buf[1500] = {0};
 
-    int nrecv = cpe_recvfrom(dgram->m_fd, buf, sizeof(buf) - 1, 0, (struct sockaddr *)&addr, &addr_len);
+    int nrecv = (int)cpe_recvfrom(dgram->m_fd, buf, sizeof(buf) - 1, 0, (struct sockaddr *)&addr, &addr_len);
     if (nrecv < 0) {
         CPE_ERROR(
             net_schedule_em(schedule),
